Call source.to_s before sending to matches_filter.
Some gems use the stylesheet helper with a symbol. For example, the
`xray-rails` gem uses this code:

```
<%= stylesheet_link_tag :xray %>
```

This would cause the underlying Sprockets code (v2.10.1) to raise
the following error: "no implicit conversion of Symbol into String".

Solve this by converting the source (:xray in the above example) to
a string before passing it to Sprockets.